add api docs

+++ database/migrations/2026_06_11_132113_create_links_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('url');
             $table->string('code', 6)->unique();
-            $table->unsignedBigInteger('clicks')->default(0);
+            $table->integer('clicks')->default(0);
             $table->timestamps();
         });
     }
